# Significantly improve performance by using a buffered writer