@Enselic 2024-10-25
Those numbers look very nice 👍 

I skimmed over the diff which is quite big. Is it possible to split this up into smaller commits/PRs to make it easier to review/handle?