@sharkdp 2024-10-25
Didn't have time to look into the code, but didn't we experiment with this a while ago and decided that we can't use a buffered writer, because we need to support interactive use cases where you run `bat`, enter a single line of input, see the output, enter the next line of input, see the output, …