@MoSal 2024-10-25
@sharkdp 
Is it an acceptable solution to flush on each read line if stdin is detected as terminal?
I just pushed a commit that does just that.