@eth-p 2024-10-28
> Is it an acceptable solution to flush on each read line if stdin is detected as terminal?

That wouldn't account for other programs piping to bat: `sed 's/A/a/' </dev/stdin | bat`

If it's possible to flush at the end of every input line regardless of the source, I would be happy with that compromise. We can always optimize it further in the future to add special exceptions for cases such as "the input source is a regular file."